chore: Move Settings.ts and SettingsTab.ts to src/config/
This matches changes made by @sytone in their fork, and will
simplify reviewing of PRs based on this layout.
