+++ src/Sort.ts
@@ -2,7 +2,7 @@ import type moment from 'moment';
 import type { Task } from './Task';
 import type { Query, SortingProperty } from './Query';
 
-import { getSettings } from './Settings';
+import { getSettings } from './config/Settings';
 
 type Comparator = (a: Task, b: Task) => number;
